requirements.txt
flask
sqlalchemy
requests
psycopg2-binary
python-dotenv

